gstreamer api refactor
