+++ py/aikit/api/pipeline.py
@@ -1,45 +1,17 @@
 import gi
 
+from py.aikit.api.data import HailoData
+
 gi.require_version('Gst', '1.0')  # define before importing Gst
 
 from gi.repository import Gst
 from py.aikit.api.commons import disable_qos
 
 import logging
-import multiprocessing
 import sys
 import setproctitle
 
 log = logging.getLogger(__name__)
-
-
-# -----------------------------------------------------------------------------------------------
-# User-defined class to be used in the callback function
-# -----------------------------------------------------------------------------------------------
-# A sample class to be used in the callback function
-# This example allows to:
-# 1. Count the number of frames
-# 2. Setup a multiprocessing queue to pass the frame to the main thread
-# Additional variables and functions can be added to this class as needed
-
-class GStreamerData(object):
-    def __init__(self):
-        self.frame_count = 0
-        self.frame_queue = multiprocessing.Queue(maxsize=3)
-        self.running = True
-
-    def increment(self):
-        self.frame_count += 1
-
-    def get_count(self):
-        return self.frame_count
-
-    def set_frame(self, frame):
-        if not self.frame_queue.full():
-            self.frame_queue.put(frame)
-
-    def get_frame(self):
-        return None if self.frame_queue.empty() else self.frame_queue.get()
 
 
 # -----------------------------------------------------------------------------------------------
@@ -56,14 +28,15 @@ class Pipeline:
         bus.connect("message", bus_call, loop)
 
     # Connect pad probe to the identity element
-    def connect_pad_probe_to_identity_element(self, on_probe_callback, shared_data) -> None:
+    def connect_pad_probe_to_identity_element(self, on_probe_callback, data: HailoData) -> None:
         identity = self.pipeline.get_by_name("identity_callback")
         if identity is None:
             log.warning("identity_callback element not found, add <identity name=identity_callback> in your pipeline where you want the callback to be called.")
         else:
             identity_pad = identity.get_static_pad("src")
-            identity_pad.add_probe(Gst.PadProbeType.BUFFER, on_probe_callback, shared_data)
+            identity_pad.add_probe(Gst.PadProbeType.BUFFER, on_probe_callback, data)
 
+    # Disable QoS to prevent frame drops
     def disable_qos(self) -> None:
         hailo_display = self.pipeline.get_by_name("hailo_display")
         if hailo_display is None:
@@ -73,7 +46,6 @@ class Pipeline:
             if xvimagesink is not None:
                 xvimagesink.set_property("qos", False)
 
-        # Disable QoS to prevent frame drops
         disable_qos(self.pipeline)
 
     def change_state_to(self, state: Gst.State):
@@ -88,11 +60,9 @@ class Pipeline:
     @staticmethod
     def __create_pipeline(pipeline_string: str, show_fps: bool):
         try:
-            # Set the process title
             setproctitle.setproctitle("Hailo Python Application")
 
-            # Initialize GStreamer
-            Gst.init(None)
+            Gst.init(None)  # Initialize GStreamer
 
             pipeline = Gst.parse_launch(pipeline_string)
